telegram-bot/requirements.txt
python-telegram-bot==20.0
appwrite==11.1.0
python-dotenv==1.1.1
fastapi==0.104.1
uvicorn==0.24.0.post1
